+++ tsconfig.json
@@ -17,7 +17,8 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "baseUrl": "./src/"
   },
   "include": ["./src/"],
   "exclude": ["node_modules/*"]
